graphvision/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "graphvision-ai"
version = "0.1.10"  
description = "Automatic Graph Classification and Data Extraction"
readme = "README.md"
authors = [{name="Aryan Gahlot"}]
license = {text = "MIT License"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch",
    "torchvision",
    "opencv-python-headless",  
    "easyocr",
    "Pillow",
    "numpy",
    "huggingface_hub",
    "ultralytics"      
]

[tool.setuptools.packages.find]
include = ["graphvision"]